Get unittest framework running

+++ lib/roken/parse_time-test.c
@@ -67,7 +67,7 @@ main(int argc, char **argv)
 	sz = unparse_time(tests[i].val, NULL, 0);
 	if  (sz != tests[i].size)
 	    errx(1, "sz (%lu) != tests[%d].size (%lu)",
-		 (unsigned long)sz, i, (unsigned long)tests[i].size);
+		 (unsigned long)sz, (int) i, (unsigned long)tests[i].size);
 
 	for (buf_sz = 0; buf_sz < tests[i].size + 2; buf_sz++) {
 
@@ -76,13 +76,13 @@ main(int argc, char **argv)
 	    sz = unparse_time(tests[i].val, buf, buf_sz);
 	    if (sz != tests[i].size)
 		errx(1, "sz (%lu) != tests[%d].size (%lu) with in size %lu",
-		     (unsigned long)sz, i,
+		     (unsigned long)sz, (int) i,
 		     (unsigned long)tests[i].size,
 		     (unsigned long)buf_sz);
 	    if (buf_sz > 0 && memcmp(buf, tests[i].str, buf_sz - 1) != 0)
 		errx(1, "test %i wrong result %s vs %s", i, buf, tests[i].str);
 	    if (buf_sz > 0 && buf[buf_sz - 1] != '\0')
-		errx(1, "test %i not zero terminated", i);
+		errx(1, "test %i not zero terminated", (int) i);
 	    rk_test_mem_free("overrun");
 
 	    buf = rk_test_mem_alloc(RK_TM_UNDERRUN, "underrun",
@@ -90,13 +90,13 @@ main(int argc, char **argv)
 	    sz = unparse_time(tests[i].val, buf, min(buf_sz, tests[i].size));
 	    if (sz != tests[i].size)
 		errx(1, "sz (%lu) != tests[%d].size (%lu) with insize %lu",
-		     (unsigned long)sz, i,
+		     (unsigned long)sz, (int) i,
 		     (unsigned long)tests[i].size,
 		     (unsigned long)buf_sz);
 	    if (buf_sz > 0 && strncmp(buf, tests[i].str, min(buf_sz, tests[i].size) - 1) != 0)
-		errx(1, "test %i wrong result %s vs %s", i, buf, tests[i].str);
+		errx(1, "test %i wrong result %s vs %s", (int) i, buf, tests[i].str);
 	    if (buf_sz > 0 && buf[min(buf_sz, tests[i].size) - 1] != '\0')
-		errx(1, "test %i not zero terminated", i);
+		errx(1, "test %i not zero terminated", (int) i);
 	    rk_test_mem_free("underrun");
 	}
 
@@ -104,14 +104,14 @@ main(int argc, char **argv)
 				tests[i].str, tests[i].size + 1);
 	j = parse_time(buf, "s");
 	if (j != tests[i].val)
-	    errx(1, "parse_time failed for test %d", i);
+	    errx(1, "parse_time failed for test %d", (int) i);
 	rk_test_mem_free("overrun");
 
 	buf = rk_test_mem_alloc(RK_TM_UNDERRUN, "underrun",
 				tests[i].str, tests[i].size + 1);
 	j = parse_time(buf, "s");
 	if (j != tests[i].val)
-	    errx(1, "parse_time failed for test %d", i);
+	    errx(1, "parse_time failed for test %d", (int) i);
 	rk_test_mem_free("underrun");
 
     }
